refactor: extract arithmetic operations into separate module and update main calculator to import functions, eliminating code duplication and improving maintainability
