+++ arithmetic.py
@@ -1,0 +1,8 @@
+def add(num1, num2):
+    return num1 + num2
+def sub(num1, num2):
+    return num1 - num2
+def div(num1, num2):
+    return num1 / num2
+def mul(num1, num2):
+    return num1 * num2
